requirements.txt
agenta
pandas
weaviate-client
openai
gdown
tqdm
ipdb